Update circulation def

+++ test/definitions/flux_is_integral_of_norm_along_curve_test.py
@@ -20,4 +20,4 @@ def test_basic_flux(test_args):
     # flux over circle of radius = 2
     curve = [2 * cos(flux_def.parameter), 2 * sin(flux_def.parameter)]
     result = flux_def.calculate_flux(field, curve, (0, 2 * pi))
-    assert float(convert_to(result, S.One).evalf(4)) == approx(float((8 * pi).evalf(4)), 0.001)
+    assert convert_to(result, S.One).evalf(4) == approx((8 * pi).evalf(4), 0.001)
